deleted:    Area.class 	deleted:    Area_MOR.class 	deleted:    Area_MOR.java 	deleted:    Area_MOR.txt 	deleted:    Array_Sum.java renamed:    Strings3.class -> Strings/Strings3.class 	renamed:    Strings3.java -> Strings/Strings3.java 	renamed:    Strings4.class -> Strings/Strings4.class 	renamed:    Strings4.java -> Strings/Strings4.java 	renamed:    Strings5.class -> Strings/Strings5.class 	renamed:    Strings5.java -> Strings/Strings5.java 	deleted:    Student.class
