--- Constructor1.java
@@ -1,17 +1,0 @@
-class Constructor1
-{
-int a;
-boolean b;
-float c;
-double d;
-char e;
-public static void main(String args[])
-{
-Constructor1 obj = new Constructor1();
-System.out.println(obj.a);
-System.out.println(obj.b);
-System.out.println(obj.c);
-System.out.println(obj.d);
-System.out.println(obj.e);
-}
-}
