+++ Constructors/Constructor2.class
