+++ Functions/Fnctions_OL.java
@@ -1,0 +1,17 @@
+public class Functions_OL
+{
+int Add(int a,int b)
+{
+int sum = a+b;
+return sum;
+}
+int Add(int a,int b,int c)
+{
+int sum = a+b+c;
+return sum;
+}
+int Add(int a,int b,int c,int d)
+{
+int sum = a+b+c+d;
+return sum;
+}
