+++ Strings/Strings1.java
@@ -1,0 +1,12 @@
+import java.util.*;
+public class Strings1
+{
+public static void main(String args[])
+{
+Scanner sc = new Scanner(System.in);
+String s;
+System.out.println("Enter the string: ");
+s = sc.nextLine();
+System.out.println("The entered string is: "+s);
+}
+}
